pyproject.toml
[tool.poetry]
name = "co-optimal"
version = "0.1.0"
description = "A FastAPI-based backend application with PostgreSQL database integration"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "co_optimal"}]

[tool.poetry.dependencies]
python = "^3.13.2"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
gunicorn = "^21.2.0"
psycopg2-binary = "^2.9.9"
pydantic-settings = "^2.1.0"
newrelic = "^9.2.0"
alembic = "^1.16.4"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 